chore(server): move WebhookSessionValidator impl to adapters/
`identity/services/session-validator.ts` defined BOTH the
`SessionValidator` interface AND its `WebhookSessionValidator`
implementation. The impl uses `WebhookClient` (an adapter dep), making
it an adapter — same shape as `WebhookContactService`.

Split:
- `identity/services/session-validator.ts` keeps the `SessionValidator`
  interface + the `SessionValidation` result type (the identity
  contract, 27 lines, no runtime deps).
- New file `adapters/webhook-session-validator.ts` holds the
  `WebhookSessionValidator` class.
- `standalone.ts`: import the impl from adapters/.

Tsc clean.
